[~] refactor search bar
