--- lab1/Theme.swift
@@ -1,17 +1,0 @@
-//
-//  Theme.swift
-//  lab1
-//
-//  Created by dezavv on 11/23/23.
-//
-
-import Foundation
-import SwiftUI
-
-extension Color {
-    static let primaryBlue = Color(red: 0.8, green: 0.88, blue: 1)
-}
-
-extension Image {
-    static let firstDoctor = Image("Doc")
-}
