+++ lab1/en.lproj/Localizable.strings
@@ -1,0 +1,8 @@
+/* 
+  Localizable.strings
+  lab1
+
+  Created by dezavv on 11/23/23.
+  
+*/
+"searchDoctor" = "Search doctor or health issue";
