+++ lab1/theme/Theme.swift
@@ -1,0 +1,23 @@
+//
+//  Theme.swift
+//  lab1
+//
+//  Created by dezavv on 11/23/23.
+//
+
+import Foundation
+import SwiftUI
+
+extension Color {
+    static let backgroundGray = Color(red: 0.98, green: 0.98, blue: 0.98)
+    static let textGray = Color(red: 0.53, green: 0.59, blue: 0.73)
+}
+
+extension Image {
+    static let firstDoctor = Image("Doc")
+    static let searchNormal = Image("search-normal")
+}
+
+extension Font {
+    static let poppinsMiddle = Font.custom("Poppins", size: 15)
+}
