+++ lab1/view/common/DoctorPhotoView.swift
