+++ localizable/en.lproj/Localizable.strings
@@ -1,0 +1,9 @@
+/* 
+  Localizable.strings
+  lab1
+
+  Created by dezavv on 11/23/23.
+  
+*/
+
+
